+++ examples/benchmark.py
@@ -30,7 +30,7 @@ def main() -> None:
     )
     benchmark: Benchmark = Benchmark([test_case])
     benchmark.run()
-    print(benchmark.to_string(floatfmt=".7f"))
+    print(benchmark.results.to_string(floatfmt=".7f"))
 
 
 if __name__ == "__main__":